+++ fighters/kirby/src/status/copy/ryu/special_n2.rs
@@ -35,8 +35,8 @@ unsafe extern "C" fn kirby_ryu_special_n2_command_pre(fighter: &mut L2CFighterCo
 
 unsafe extern "C" fn kirby_ryu_special_n2_command_init(fighter: &mut L2CFighterCommon) -> L2CValue {
     WorkModule::on_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_FLAG_COMMAND);
-    WorkModule::set_int64(fighter.module_accessor, hash40("ryu_special_n2") as i64, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_N_INT_MOTION_KIND);
-    WorkModule::set_int64(fighter.module_accessor, hash40("ryu_special_air_n2") as i64, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_N_INT_MOTION_KIND_AIR);
+    WorkModule::set_int64(fighter.module_accessor, hash40("ryu_special_n2"), *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_N_INT_MOTION_KIND);
+    WorkModule::set_int64(fighter.module_accessor, hash40("ryu_special_air_n2"), *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_N_INT_MOTION_KIND_AIR);
     let speed_x = KineticModule::get_sum_speed_x(fighter.module_accessor, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_MAIN);
     let mut speed_y = KineticModule::get_sum_speed_y(fighter.module_accessor, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_MAIN);
     if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
