+++ fighters/lucario/src/status/special_n.rs
@@ -3,8 +3,8 @@ use super::super::helper::*;
 
 unsafe extern "C" fn lucario_special_n_main(fighter: &mut L2CFighterCommon) -> L2CValue {
     WorkModule::off_flag(fighter.module_accessor, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_FLAG_MOT_INHERIT);
-    WorkModule::set_int64(fighter.module_accessor, hash40("special_n_start") as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
-    WorkModule::set_int64(fighter.module_accessor, hash40("special_air_n_start") as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
+    WorkModule::set_int64(fighter.module_accessor, hash40("special_n_start"), *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
+    WorkModule::set_int64(fighter.module_accessor, hash40("special_air_n_start"), *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
     WorkModule::set_customize_no(fighter.module_accessor, 0, 0);
     // lucario_drain_aura(fighter, false);
     lucario_special_set_kinetic(fighter);
@@ -57,8 +57,8 @@ unsafe extern "C" fn lucario_special_n_hold_main(fighter: &mut L2CFighterCommon)
         ground_mot = hash40("special_n_hold");
         air_mot = hash40("special_air_n_hold");
     }
-    WorkModule::set_int64(fighter.module_accessor, ground_mot as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
-    WorkModule::set_int64(fighter.module_accessor, air_mot as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
+    WorkModule::set_int64(fighter.module_accessor, ground_mot, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
+    WorkModule::set_int64(fighter.module_accessor, air_mot, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
     ArticleModule::change_status(
         fighter.module_accessor,
         *FIGHTER_LUCARIO_GENERATE_ARTICLE_AURABALL,
@@ -144,8 +144,8 @@ unsafe extern "C" fn lucario_special_n_shoot_main(fighter: &mut L2CFighterCommon
         ground_mot = hash40("special_n_shoot");
         air_mot = hash40("special_air_n_shoot");
     }
-    WorkModule::set_int64(fighter.module_accessor, ground_mot as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
-    WorkModule::set_int64(fighter.module_accessor, air_mot as i64, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
+    WorkModule::set_int64(fighter.module_accessor, ground_mot, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_GROUND_MOT);
+    WorkModule::set_int64(fighter.module_accessor, air_mot, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AIR_MOT);
     let charge = WorkModule::get_int(fighter.module_accessor, *FIGHTER_LUCARIO_INSTANCE_WORK_ID_INT_AURABALL_CHARGE_FRAME);
     let max_charge_frame = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_n"), hash40("max_charge_frame"));
     if max_charge_frame < charge as f32 {
